remove Agent interface, rename file
