+++ backend/src/services/agent-services/create-transaction-from-text-service.test.ts
@@ -3,7 +3,7 @@ import { beforeEach, describe, expect, it, jest } from "@jest/globals";
 import { AIMessage, ReactAgent, ToolMessage } from "langchain";
 import { CREATE_TRANSACTION_TOOL_NAME } from "../../langchain/tools/create-transaction";
 import { fakeTransaction } from "../../utils/test-utils/models/transaction-fakes";
-import { AgentTraceMessageType } from "../ports/agent";
+import { AgentTraceMessageType } from "../ports/agent-types";
 import { TransactionService } from "../transaction-service";
 import { CreateTransactionFromTextService } from "./create-transaction-from-text-service";
