+++ backend/src/services/agent-services/create-transaction-from-text-service.ts
@@ -9,7 +9,7 @@ import {
 import { Transaction } from "../../models/transaction";
 import { Failure, Result, Success } from "../../types/result";
 import { formatDateAsYYYYMMDD } from "../../utils/date";
-import { AgentTraceMessage } from "../ports/agent";
+import { AgentTraceMessage } from "../ports/agent-types";
 import { TransactionService } from "../transaction-service";
 
 const createdTransactionSchema = z.discriminatedUnion("success", [
